feat：支持管理员查看项目成员 #9620

+++ src/backend/ci/core/auth/biz-auth/src/main/kotlin/com/tencent/devops/auth/provider/rbac/service/migrate/AbMigratePolicyService.kt
@@ -157,7 +157,7 @@ abstract class AbMigratePolicyService(
                 projectName = groupInfo.resourceName
             )
             authorizationScopeList.forEach { authorizationScope ->
-                v2ManagerService.grantRoleGroupV2(groupInfo.relationId.toInt(), authorizationScope)
+                v2ManagerService.grantRoleGroupV2(groupInfo.relationId, authorizationScope)
             }
         }
     }
